fix(preload): split channel constants to avoid zod evaluation chain (PR B)
修复发起人在 W1 抽查 pull main 后第一次 pnpm dev 撞到的 preload sandbox 白屏。
对齐 03-architecture.md §三 "contextBridge + typed IPC，不用 remote" 的
thin-preload 精神，并提前给 TD-002 (#30) 一份 preload bundle 减重收益。

## 根因

preload 在 sandbox:true 下不能 runtime require external module。M0 留下的
preload 从 @opentrad/shared 根 export 拿 IpcChannels（value import），会
触发 ipc.ts 顶部的 `import { z } from "zod"` evaluation chain → vite 在
preload bundle 里留 require("zod") → sandbox 加载 preload 时 module not
found → 窗口白屏 + window.api 未定义。

W1 之前没暴露是因为 M0 渲染端只用了 `cc.status()`，preload 跑通的 happy
path 浅。M1 #19 / #20 开始 zod schemas 越来越多，evaluation chain 越来越
重，第一次有人本地真把 dev 跑起来时（发起人抽查）就炸。

## 改动

### packages/shared/src/channels.ts（新增）

把所有纯 const channel 名独立出来，**本文件不 import zod**。

### packages/shared/src/types/ipc.ts

删除 IpcChannels 定义，改成 `export { IpcChannels } from "../channels"`，
保持向后兼容（main 进程 / packages 仍可从根 export 拿，零迁移）。

### packages/shared/src/index.ts

加 `export * from "./channels"`，shared root 仍暴露 IpcChannels。

### packages/shared/package.json

加 `exports` 字段暴露 `@opentrad/shared/channels` 子路径，
让 preload 直接 import 子路径。

### apps/desktop/src/preload/index.ts

`import { IpcChannels } from "@opentrad/shared"` →
`import { IpcChannels } from "@opentrad/shared/channels"`。
顶部加注释，明确 preload 永远只能从 channels 子路径拿运行时常量
（type-only import 编译时擦除，从根 export 拿没问题）。

## 验证

- `pnpm typecheck`：8 packages 全过
- `pnpm test`：154 tests 全过
- `pnpm lint`：clean
- **`pnpm exec electron-vite build`**：
  * preload bundle 18.73 KB → **2.32 KB**（-87%）
  * `grep "require(\"zod\")\|from\"zod\"" out/preload/index.js` → **0 匹配**
  * 82 行 bundle，纯 IPC channel 转发

## 边界原则

preload 永远只 import @opentrad/shared/channels（或 type-only import）。
preload 不做 zod 校验；所有 zod 校验在 main 进程的 IPC handlers 里做。
未来贡献者新加 IPC 时，channels.ts 加常量、ipc.ts 加 schema、preload 加
转发，三步分离不互相污染。

## TD-002 (#30) 顺手收益

retrospective-m0.md §五 TD-002 原计划 M1 中期把 preload bundle 138 KB
拆分。本 PR 把 preload 减到 2.32 KB（远优于 ≤ 80 KB 目标），#30 这部分
基本提前完成；#30 剩下 IPC channels 按 domain 拆分（cc / pty / session
等多个文件）的工作量大幅降低。

## Refs

- 修发起人在 W1 抽查发现的 bug 2
- 配套 PR A：[#34](https://github.com/open-trad/opentrad/pull/34)（已合，bug 1 electron ABI）
- 解锁 #25 重启（rebase main 后 IPC channels 层不再受 zod 链污染）
